requirements.txt
python-telegram-bot==21.7
pdfplumber==0.11.4
httpx==0.27.2
python-dotenv==1.0.1
rapidfuzz==3.10.1
